requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
requests
python-dotenv
twilio
fastapi
uvicorn
Faker